requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pillow==10.3.0
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
opencv-python-headless==4.9.0.80
torchxrayvision==1.3.2
scikit-learn==1.4.2
scipy==1.13.0
psycopg2-binary==2.9.9
aiofiles==23.2.1
matplotlib==3.8.4
scikit-image==0.23.2
grad-cam==1.5.5
requests==2.31.0
httpx
